Cambios en UpdatesProvider
- Se cambiaron todos los identificadores de Integer a Long.
- Se agregaron y modificaron métodos.
- Se modificó el acceso a la base de datos apropiadamente.

+++ src/main/java/com/mypersonalupdates/db/actions/UpdatesProviderActions.java
@@ -1,5 +1,6 @@
 package com.mypersonalupdates.db.actions;
 
+import com.mypersonalupdates.providers.UpdatesProvider;
 import com.mypersonalupdates.db.mappers.UpdatesProviderAttributeMapper;
 import com.mypersonalupdates.providers.UpdatesProviderAttribute;
 import org.skife.jdbi.v2.sqlobject.Bind;
@@ -8,20 +9,23 @@ import org.skife.jdbi.v2.sqlobject.customizers.Mapper;
 
 import java.util.List;
 
+/**
+ * Acciones en la base de datos para la clase {@link UpdatesProvider}
+ */
 public interface UpdatesProviderActions {
     @SqlQuery("SELECT name FROM updates_provider WHERE ID = :id")
     String getName(
-            @Bind("id") Integer id
+            @Bind("id") Long id
     );
 
     @SqlQuery("SELECT description FROM updates_provider WHERE ID = :id")
     String getDescription(
-            @Bind("id") Integer id
+            @Bind("id") Long id
     );
 
     @SqlQuery("SELECT providerID, attrID FROM updates_provider_attribute_association WHERE providerID = :id")
     @Mapper(UpdatesProviderAttributeMapper.class)
     List<UpdatesProviderAttribute> getAttributes(
-            @Bind("id") Integer id
+            @Bind("id") Long id
     );
 }
